finish extraction of expectations tool & package
